Added a lot of things

+++ C++/ProcessManager/Process.cpp
@@ -3,17 +3,28 @@
 #include "Process.h"
 #include <exception>
 
+#include "CriticalSectionGuard.h"
+
+#include <winternl.h>
+
+#include <iostream>
+
+#pragma comment(lib, "Ntdll")
 
 
 Process::Process(const ustring &pCmdLine) : mCmdLine(pCmdLine)
 {
-	mLogger = NULL;
+	mLogger = nullptr;
+	hRegisterWait = nullptr;
+	hRegisterWaitMonitor = nullptr;
 	initInfo();
+	isActive = false;
+	isMonitoring = false;
 }
 
 Process::Process(const ustring &pCmdLine, AbstractLogger *logger) : Process(pCmdLine)
 {
-	mLogger = logger;
+	InterlockedExchangePointer(reinterpret_cast<volatile PVOID*>(&mLogger), logger);
 }
 
 
@@ -22,51 +33,69 @@ Process::~Process()
 	clearInfo();
 }
 
-
+// Creates process and calls monitor method. Throws exception if fails to start new process, 
+// if process is already running, logs info about it.
 void Process::run()
 {
-	clearInfo();
-	initInfo();
-	
-	// Start the child process. 
-	if (!CreateProcess(NULL,   // No module name (use command line)
-		_tcsdup(mCmdLine.c_str()),        // Command line
-		NULL,           // Process handle not inheritable
-		NULL,           // Thread handle not inheritable
-		FALSE,          // Set handle inheritance to FALSE
-		0,              // No creation flags
-		NULL,           // Use parent's environment block
-		_tcsdup(TEXT("C:\\Windows\\Temp")),
-		&mStartupInfo,            // Pointer to STARTUPINFO structure
-		&mProcessInfo)           // Pointer to PROCESS_INFORMATION structure
-		)
+	CriticalSectionGuard autoLock(&mLock);
+	if (!isActive)
 	{
-		throw std::runtime_error(std::string("CreateProcess failed, errcode: ") + std::to_string(GetLastError()) + "\n");
+		clearInfo();
+		initInfo();
+
+		if (mCmdLine.empty())
+			throw std::runtime_error((std::string("Cmdline empty") + std::to_string(GetLastError())) + std::string("\n"));
+
+		// Start the child process. 
+		if (!CreateProcess(NULL,   // No module name (use command line)
+			_tcsdup(mCmdLine.c_str()),        // Command line
+			NULL,           // Process handle not inheritable
+			NULL,           // Thread handle not inheritable
+			FALSE,          // Set handle inheritance to FALSE
+			0,              // No creation flags
+			NULL,           // Use parent's environment block
+			_tcsdup(TEXT("C:\\Windows\\Temp")),
+			&mStartupInfo,            // Pointer to STARTUPINFO structure
+			&mProcessInfo)           // Pointer to PROCESS_INFORMATION structure
+			)
+		{
+			throw std::runtime_error(std::string("CreateProcess failed, errcode: ") + std::to_string(GetLastError()) + "\n");
+		}
+		if (mLogger)
+			mLogger->logData(StringBuilder() << TEXT("[LOG] ") << TEXT("Process has started, PID ")
+											 << TO_USTRING(mProcessInfo.dwProcessId) << TEXT("\n"));
+
+		onEvent(EventType::Started);
+		isActive = true;
+
+		if (!RegisterWaitForSingleObject(&hRegisterWait, mProcessInfo.hProcess, onTerminated, static_cast<void*>(this), INFINITE, WT_EXECUTEONLYONCE))
+		{
+			throw std::runtime_error((std::string("RegisterWaitForSingleObject failed, errcode: ") + std::to_string(GetLastError())) + std::string("\n"));
+		}
+
 	}
-	if (mLogger)
-		mLogger->logData(StringBuilder() << TEXT("[LOG] ") << TEXT("Process has started, PID ") 
-										 << TO_USTRING(mProcessInfo.dwProcessId) << TEXT("\n"));
-
-	onEvent(EventType::Started);
-
-	monitor(mProcessInfo.dwProcessId);
+	else
+	{
+		if (mLogger)
+			mLogger->logData(StringBuilder() << TEXT("[LOG] Process is already running, PID ")
+											 << TO_USTRING(mProcessInfo.dwProcessId) << TEXT("\n"));
+	}
 }
 
 void Process::terminate()
 {
-	if (!TerminateProcess(mProcessInfo.hProcess, 0))
-	{
-		throw std::runtime_error(std::string("TerminateProcess failed, errcode: ") + std::to_string(GetLastError()) + "\n");
-	}
+	CriticalSectionGuard autoLock(&mLock);
+	if (mProcessInfo.hProcess)
+		if (!TerminateProcess(mProcessInfo.hProcess, 0))
+			throw std::runtime_error(std::string("TerminateProcess failed, errcode: ") + std::to_string(GetLastError()) + "\n");
 }
 
 void Process::restart()
 {
-	if (isActive())
+	if (isActive)
 	{
 		terminate();
 	}
-	//Sleep(10);
 	run();
 }
 
@@ -87,39 +116,31 @@ void Process::clearInfo()
 
 int Process::processId() const
 {
-	return GetProcessId(mProcessInfo.hProcess);
-}
-
-bool Process::isActive() const
-{
-	DWORD exitCode = 0;
-	if (GetExitCodeProcess(mProcessInfo.hProcess, &exitCode) == 0)
-	{
-		throw std::runtime_error((std::string("GetExitCodeProcess failed, errcode: ") + std::to_string(GetLastError())) + std::string("\n"));
-	}
-	return exitCode == STILL_ACTIVE ? true : false;
+	return mProcessInfo.dwProcessId;
 }
 
 
 void Process::setEventCallback(std::function<void()> callback, const EventType evenType)
 {
-	mCallbaks[evenType] = callback;
+	mCallbacks[evenType] = callback;
 }
 
-
+// Static function that calls when process terminates. It decides what to log and what callbacks to call.
 void NTAPI Process::onTerminated(PVOID lpParameter, BOOLEAN TimerOrWaitFired)
 {
-	const Process *sender = static_cast<Process*>(lpParameter); // sender of event
+	
+	Process *sender = static_cast<Process*>(lpParameter); // sender of event
 	if (sender)
 	{
 		DWORD exitCode = 0;
+		
 		if (GetExitCodeProcess(sender->mProcessInfo.hProcess, &exitCode))
 		{
 			if (exitCode == 0)
 			{
 				if (sender->mLogger)
 					sender->mLogger->logData(StringBuilder() << TEXT("[LOG] ") << TEXT("Process has been stoped, PID ")
-					<< TO_USTRING(sender->mProcessInfo.dwProcessId) << TEXT("\n"));
+															 << TO_USTRING(sender->mProcessInfo.dwProcessId) << TEXT("\n"));
 
 				sender->onEvent(EventType::Stoped);
 			}
@@ -127,54 +148,126 @@ void NTAPI Process::onTerminated(PVOID lpParameter, BOOLEAN TimerOrWaitFired)
 			{
 				if (sender->mLogger)
 					sender->mLogger->logData(StringBuilder() << TEXT("[LOG] ") << TEXT("Process has crashed, PID ")
-					<< TO_USTRING(sender->mProcessInfo.dwProcessId) << TEXT("\n"));
-
+															 << TO_USTRING(sender->mProcessInfo.dwProcessId) << TEXT("\n"));
 
 				sender->onEvent(EventType::Crashed);
 			}
 		}
+
+		UnregisterWait(sender->hRegisterWait);
+		sender->hRegisterWait = nullptr;
+		sender->isActive = false;
+		if (sender->isMonitoring)
+		{
+			sender->run();
+		}
+
 		
 	}
 	
 }
 
-void Process::setLogger(AbstractLogger *logger)
+ustring NTAPI Process::processCommandLine(void *handle)
 {
-	mLogger = logger;
-}
+	PROCESS_BASIC_INFORMATION  pInfo;
+	PVOID rtlUserProcParamsAddress;
+	UNICODE_STRING commandLine;
+	WCHAR *commandLineContents;
 
-void Process::monitor(int pId)
-{
-	HANDLE hwndProc = OpenProcess(PROCESS_ALL_ACCESS, FALSE, pId);
-	ustring infoString;
+	LONG status = NtQueryInformationProcess(handle,
+		ProcessBasicInformation,
+		&pInfo,
+		sizeof(pInfo),
+		NULL);
 
-	if (hwndProc == NULL)
+	bool check;
+
+	void *pedAddress = pInfo.PebBaseAddress;
+	/* get the address of ProcessParameters */
+	if (!ReadProcessMemory(handle, (PCHAR)pedAddress + 0x10,
+		&rtlUserProcParamsAddress, sizeof(PVOID), NULL))
 	{
-		infoString = StringBuilder() << TEXT("[LOG] ") << TEXT("Couldn`t open a process to monitor, PID: ") 
-									 << TO_USTRING(pId) << TEXT("\n");
-
-		mLogger->logData(infoString);
-		return;
+		check = false;
 	}
 
-	infoString = StringBuilder() << TEXT("[LOG] ") << TEXT("Opened a process to monitor, PID: ") 
-								 << TO_USTRING(pId) << TEXT("\n");
-
-	mLogger->logData(infoString);
-	
-	void *hRegisterWait = NULL;
-	CloseHandle(mProcessInfo.hProcess); // Close previous handle
-	mProcessInfo.hProcess = hwndProc;
-	mProcessInfo.dwProcessId = pId;
-
-	if (!RegisterWaitForSingleObject(&hRegisterWait, hwndProc, onTerminated, static_cast<void*>(this), INFINITE, WT_EXECUTEONLYONCE))
+	/* read the CommandLine UNICODE_STRING structure */
+	if (!ReadProcessMemory(handle, (PCHAR)rtlUserProcParamsAddress + 0x40,
+		&commandLine, sizeof(commandLine), NULL))
 	{
-		throw std::runtime_error((std::string("RegisterWaitForSingleObject failed, errcode: ") + std::to_string(GetLastError())) + std::string("\n"));
+		check = false;
+	}
+
+	/* allocate memory to hold the command line */
+	commandLineContents = (WCHAR *)malloc(commandLine.Length);
+
+	/* read the command line */
+	if (!ReadProcessMemory(handle, commandLine.Buffer,
+		commandLineContents, commandLine.Length, NULL))
+	{
+		check = false;
+	}
+
+	return ustring(check ? commandLineContents : TEXT(""));
+}
+
+void Process::setLogger(AbstractLogger *logger)
+{
+	InterlockedExchangePointer(reinterpret_cast<volatile PVOID*>(&mLogger), logger);
+}
+
+// Function that takes pID and starts to monitor selected process.
+// Can be used in method run or separatly
+void Process::monitor(const int pId)
+{	
+	CriticalSectionGuard autoLock(&mLock);
+	if (!hRegisterWaitMonitor)
+	{
+		HANDLE hwndProc = OpenProcess(PROCESS_ALL_ACCESS, FALSE, pId);
+		ustring infoString;
+
+		if (!hwndProc)
+		{
+			infoString = StringBuilder() << TEXT("[LOG] ") << TEXT("Couldn`t open a process to monitor, PID: ")
+				<< TO_USTRING(pId) << TEXT("\n");
+
+			if (mLogger)
+				mLogger->logData(infoString);
+
+			return;
+		}
+
+		infoString = StringBuilder() << TEXT("[LOG] ") << TEXT("Opened a process to monitor, PID: ")
+			<< TO_USTRING(pId) << TEXT("\n");
+
+		mLogger->logData(infoString);
+
+		clearInfo(); // Close handles which were used before. Because we will use handle that we got from OpenProcess function.
+		mProcessInfo.hProcess = hwndProc;
+		mProcessInfo.dwProcessId = pId;
+
+		mCmdLine = processCommandLine(mProcessInfo.hProcess);
+
+		if (!RegisterWaitForSingleObject(&hRegisterWaitMonitor, hwndProc, onTerminated, static_cast<void*>(this), INFINITE, WT_EXECUTEONLYONCE))
+		{
+			throw std::runtime_error((std::string("RegisterWaitForSingleObject failed, errcode: ") + std::to_string(GetLastError())) + std::string("\n"));
+		}
+		isMonitoring = true;
+	}
+	else
+	{
+		if (mLogger)
+			mLogger->logData(StringBuilder() << TEXT("[LOG] Process is already monitoring, PID ")
+			<< TO_USTRING(mProcessInfo.dwProcessId) << TEXT("\n"));
 	}
 }
 
 void Process::onEvent(const EventType type) const
 {
-	if (!mCallbaks[type]._Empty())
-		mCallbaks[type]();
+	if (!mCallbacks[type]._Empty())
+		mCallbacks[type]();
+}
+
+bool Process::isRunning() const
+{
+	return isActive | isMonitoring;
 }
